move setup_uid_manager() to mockbuild.uid module
This will ease usage by thirdparties. E.g., fedora-review

The motivation is that fedora-review calls load_config without uidManager param
which fatally fails because recent changes require the uidManager there and it
cannot be None.
After this change they call setup_uid_manager() easily.
