+++ mock/py/mockbuild/uid.py
@@ -6,6 +6,7 @@
 
 import ctypes
 import errno
+import grp
 import os
 import pwd
 
@@ -13,6 +14,28 @@ from .trace_decorator import traceLog
 
 _libc = ctypes.CDLL(None, use_errno=True)
 
+
+@traceLog()
+def setup_uid_manager(mockgid):
+    unprivUid = os.getuid()
+    unprivGid = os.getgid()
+
+    # sudo
+    if os.environ.get("SUDO_UID") is not None:
+        unprivUid = int(os.environ['SUDO_UID'])
+        os.setgroups((mockgid,))
+        unprivGid = int(os.environ['SUDO_GID'])
+
+    # consolehelper
+    if os.environ.get("USERHELPER_UID") is not None:
+        unprivUid = int(os.environ['USERHELPER_UID'])
+        unprivName = pwd.getpwuid(unprivUid).pw_name
+        secondary_groups = [g.gr_gid for g in grp.getgrall() if unprivName in g.gr_mem]
+        os.setgroups([mockgid] + secondary_groups)
+        unprivGid = pwd.getpwuid(unprivUid)[3]
+
+    uidManager = UidManager(unprivUid, unprivGid)
+    return uidManager
 
 class UidManager(object):
     @traceLog()
